refactor: Keep the case-folded probe out of match-mapping
match-mapping.ts holds the single rule that turns a file into the import
specifier it is shared under, so that the reachability walk and the
resolveGlob expansion cannot disagree. A second, case-folding matcher
exported from there reads as an alternative rule even though it only ever
feeds a warning and never decides whether a mapping is used.

It has one caller. Moving it next to that caller keeps the invariant
legible and puts the probe beside the message it explains.
